api/website/requirements.txt
fastapi[all]
openai
scrapy
nltk
python-dotenv
pydantic
beautifulsoup4
langchain